server/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
psycopg2-binary==2.9.6
python-dotenv==1.0.0
gunicorn==20.1.0